refactor(node): split PeerNode helpers
Extract load tracking, noderef parsing, and offer sending\ninto dedicated helpers to reduce PeerNode dependencies.
