+++ src/main/java/network/crypta/node/BaseSender.java
@@ -14,9 +14,9 @@ import network.crypta.io.comm.NotConnectedException;
 import network.crypta.keys.Key;
 import network.crypta.keys.NodeCHK;
 import network.crypta.keys.NodeSSK;
-import network.crypta.node.PeerNode.RequestLikelyAcceptedState;
-import network.crypta.node.PeerNode.SlotWaiter;
-import network.crypta.node.PeerNode.SlotWaiterFailedException;
+import network.crypta.node.PeerNodeLoadTracker.RequestLikelyAcceptedState;
+import network.crypta.node.PeerNodeLoadTracker.SlotWaiter;
+import network.crypta.node.PeerNodeLoadTracker.SlotWaiterFailedException;
 import network.crypta.support.TimeUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
